+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updating of calendars.
 * Allow for etag check to be used on event deletions.
 * Finding of single events and calendars by URLs.
+* Check if calendar creation is possible.
 
 ## [0.0.1] - 2021-06-13